:hammer: Refactor the datasets folder & :white_check_mark: Fix loss test
