+++ torch_uncertainty/datamodules/cifar10.py
@@ -10,7 +10,8 @@ from torch import nn
 from torch.utils.data import DataLoader, Dataset, random_split
 from torchvision.datasets import CIFAR10, SVHN
 
-from ..datasets import CIFAR10C, CIFAR10H, AggregatedDataset
+from ..datasets import AggregatedDataset
+from ..datasets.classification import CIFAR10C, CIFAR10H
 from ..transforms import Cutout
