+++ torch_uncertainty/datamodules/imagenet.py
@@ -11,7 +11,7 @@ from torch import nn
 from torch.utils.data import DataLoader, Dataset
 from torchvision.datasets import DTD, SVHN, ImageNet, INaturalist
 
-from ..datasets import ImageNetA, ImageNetO, ImageNetR
+from ..datasets.classification import ImageNetA, ImageNetO, ImageNetR
 
 
 # fmt: on
