+++ torch_uncertainty/datasets/__init__.py
@@ -1,14 +1,2 @@
 # flake8: noqa
 from .aggregated_dataset import AggregatedDataset
-from .cifar import CIFAR10C, CIFAR10H, CIFAR10N, CIFAR100C, CIFAR100N
-from .fractals import Fractals
-from .imagenet import (
-    ImageNetA,
-    ImageNetO,
-    ImageNetR,
-    TinyImageNet,
-    TinyImageNetC,
-)
-from .mnist_c import MNISTC
-from .not_mnist import NotMNIST
-from .uci_regression import UCIRegression
