+++ torch_uncertainty/datasets/aggregated_dataset.py
@@ -1,8 +1,18 @@
+# fmt: off
 import torch
 from torch.utils.data import Dataset
 
 
+# fmt: on
 class AggregatedDataset(Dataset):
+    """Virtually interlace multiple copies of a dataset to train ensembles with
+    different batch orders.
+
+    Args:
+        dataset: The dataset to be interlaced.
+        n_dataloaders: The number of dataloaders to be used for training.
+    """
+
     def __init__(self, dataset: Dataset, n_dataloaders: int) -> None:
         super().__init__()
         self.dataset = dataset
