+++ torch_uncertainty/datasets/classification/__init__.py
@@ -1,0 +1,12 @@
+# flake8: noqa
+from .cifar import CIFAR10C, CIFAR10H, CIFAR10N, CIFAR100C, CIFAR100N
+from .fractals import Fractals
+from .imagenet import (
+    ImageNetA,
+    ImageNetO,
+    ImageNetR,
+    TinyImageNet,
+    TinyImageNetC,
+)
+from .mnist_c import MNISTC
+from .not_mnist import NotMNIST
