+++ torch_uncertainty/datasets/classification/imagenet/base.py
