+++ torch_uncertainty/datasets/classification/imagenet/tiny_imagenet.py
